feat(PlantReg): Shifting the purpose of the InternalApiService.cs for bridge and the new console app.
