+++ ProjectDaedalus/ProjectDaeadalus.Bridge/Services/InternalApiService.cs
@@ -28,21 +28,6 @@ namespace ProjectDaeadalus.Bridge.Services
             var responseJson = await response.Content.ReadAsStringAsync();
             return JsonSerializer.Deserialize<T>(responseJson);
         }
-
-        public async Task<T> GetAsync<T>(string endpoint)
-        {
-            var response = await _httpClient.GetAsync($"{_baseUrl}/{endpoint}");
-            response.EnsureSuccessStatusCode();
-        
-            var responseJson = await response.Content.ReadAsStringAsync();
-            return JsonSerializer.Deserialize<T>(responseJson);
-        }
-
-        public async Task<bool> DeleteAsync(string endpoint)
-        {
-            var response = await _httpClient.DeleteAsync($"{_baseUrl}/{endpoint}");
-            return response.IsSuccessStatusCode;
-        }
     }
 }
