+++ ProjectDaedalus/ProjectDaedalus.Scripts/Program.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace ProjectDaeadalus.Scripts
+{
+    class Program
+    {
+        static async Task Main(string[] args)
+        {
+            
+        }
+    }
+}
