+++ ProjectDaedalus/ProjectDaedalus.Scripts/Services/IInternalApiService.cs
@@ -1,0 +1,8 @@
+namespace ProjectDaeadalus.Scripts.Services;
+
+public interface IInternalApiService
+{
+    Task<T> PostManyAsync<T>(string endpoint, IEnumerable<object> data);
+    Task<T> GetAsync<T>(string endpoint);
+    Task<bool> DeleteAsync(string endpoint);
+}
